rename home to chat & add timestamp
